Moved encode/decode steps into Talk UI
The Transceiver really should only be sending and receiving exactly what
it's told.  If the Transceiver were a real radio, it probably wouldn't
do any encoding or decoding directly.  Either the operator would be
supply the encoded message, or there would be an additional piece of
equipment performing the encoding/decoding.  This change aligns the
program with the real world